Dodałem sprawozdania

+++ Czujniki_srodowiska/app/src/main/java/pl/lodz/p/imsi/maciekka/czujniki_srodowiska/Czujniki.java
@@ -65,8 +65,8 @@ public class Czujniki extends AppCompatActivity implements SensorEventListener {
             w_wilgotnosc.setText(String.format("Wartość wilgotności wzglednej = %.2f", wilgotnosc_w_procentach));
         }
 //        if (event.sensor.getType()==Sensor.TYPE_RELATIVE_HUMIDITY && event.sensor.getType()==Sensor.TYPE_AMBIENT_TEMPERATURE) {
-            punkt_rosy = (float)(Math.pow(wilgotnosc_w_procentach / 100, 1 / 8) *
-                    ((Math.abs((112 + (0.9 * temperatura_w_stopniach)))) + (0.1 * temperatura_w_stopniach) - 112));
+            punkt_rosy = (float)Math.pow(wilgotnosc_w_procentach / 100, 1 / 8) *
+                    (float)Math.abs((112 + (0.9 * temperatura_w_stopniach))) + (float)(0.1 * temperatura_w_stopniach) - (float)112;
             bezwzgledna_wilgotnosc=(float)(wilgotnosc_w_procentach/punkt_rosy);
             EditText w_punkt_rosy = (EditText) findViewById(R.id.punkt_rosy);
             w_punkt_rosy.setText(String.format("Wartość punktu rosy = %.2f", punkt_rosy));
